backend/requirements.txt
 fastapi
 uvicorn
 grpcio
 grpcio-tools
 asyncpg
 aiokafka
 httpx